+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Allow case arm statements to be formatted inline.
+
 ## [0.4.0-rc1] - 2025-02-20
 
 ### Fixed